requirements.txt
absl-py==0.9.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
cmake==3.16.3
cycler==0.10.0
Cython==0.29.16
future==0.18.2
fvcore==0.1.dev200420
google-auth==1.14.1
google-auth-oauthlib==0.4.1
grpcio==1.28.1
idna==2.9
intel-openmp==2020.0.133
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.1.1
mkl==2019.0
mkl-include==2019.0
mock==4.0.2
ninja==1.9.0.post1
numpy==1.17.0
oauthlib==3.1.0
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
pandas==1.0.1
Pillow==6.1.0
portalocker==1.7.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pycparser==2.20
pydot==1.4.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.3
six==1.14.0
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
torch==1.4.0
torchsummary==1.5.1
torchvision==0.5.0
tqdm==4.35.0
typing==3.7.4.1
urllib3==1.25.9
Werkzeug==1.0.1
yacs==0.1.6